HTTP Server is able to communicate with socket server and request the last image of a device.
When a device connects to the socket server it is registered in the
DeviceRegistry.
The HTTP server will query the DeviceRegistry by owner name and will
look in the list of devices by the device name. The image data is
retreived from device object and used as response of the request.